Move v1 samples to v1 instead of deleting them
