+++ custom-graders/DemoJavaGraderWithNetwork/Dockerfile
