+++ custom-graders/DemoJavaGraderWithRichFeedback/Dockerfile
