+++ custom-graders/DemoPythonGrader/README.md
@@ -10,25 +10,27 @@ At a high-level, this custom grader accomplishes autograding of student submissi
 3. The grader then checks for the student submission in the `/shared/submission/` directory. This is the standard submission directory across all courses on Coursera, and is a read-only directory.
 4. After checking that the submission has a `.py` extension, test cases are randomly generated according to the number of test cases defined in Step 2. Corresponding solutions are also generated using the files in `/autograder/solutions/`.
 5. The grader then checks for a matching partId and runs the test cases through the learner submission.
-6. After comparing the learner submission outputs to the solutions, the grader calculates the total score and sends back a JSON object with `fractionalScore` and `feedback` fields. On the Coursera platform, this JSON object can be viewed by the learner; however, the learner does not have access to `stderr`. It is important to include appropriate feedback from the custom grader into the `feedback` field of this JSON object.
+6. After comparing the learner submission outputs to the solutions, the grader calculates the total score and writes to `/shared/feedback.json` with `fractionalScore` and `feedback` fields. On the Coursera platform, this feedback can be viewed by the learner; however, the learner does not have access to container logs (including both stdout and stderr logs) -- only course admin and staff do. Thus, it is important to include appropriate feedback from the custom grader into the `feedback` field of this JSON object. 
 
 # File list and descriptions
 The student is asked to create two Python functions: (1) "Negative", which takes a single numeric input,  flips the sign of the value, and outputs the result, and (2) "Multiply", which takes a list of numbers and outputs the product of all numbers in that list. The corresponding `partId`s are `5ShhY` for "Negative" and `Jn4Lh` for "Multiply". You can find the solution files in `/sample-submissions/`.
 <br><br>
-The following files are required to build a grader image:
+The following files, all in `/autograder`, are required to build a grader image. You can navigate into the `/autograder` folder, and compress all of these into a .zip file for platform upload. Make sure that the `Dockerfile` is at the root directory.
 * `Dockerfile` sets up the Docker image. It includes a Python 3 environment and copies over the relevant files for grading.
 * `grader.py` is the main grading file and serves as the `ENTRYPOINT` for the Docker image.
 * `testCases.py` is the file that contains `partId`'s as well as the logic for assigning the test cases for each part of the assignment. `grader.py` uses this for generating test cases and the correct output cases to compare to the learner submission.
 * `util.py` contains helper functions that are used in `grader.py`.
 * `/solutions/` directory contains the solution files that are used to generate correct outputs to the test cases.
 
-# Instructions
-1. Run the below command to build the Docker image with the tag `autograder`:
-<br>`$ docker build -t autograder autograder/`
+The folder `/sample-submissions` is not required for successful build, but in this template, is provided for your convenience for local testing with `coursera-autograder`. See below.
 
-2. Run the grader locally using [`courseraprogramming`](https://github.com/coursera/courseraprogramming) and provided sample submissions in the `sample-submissions` folder:
-<br>`$ courseraprogramming grade local autograder sample-submissions/sample-Negative/ partId 5ShhY`
+# Instructions
+1. Run the below command to build the Docker image with the image tag `pythongrader`:
+<br>`$ docker build -t pythongrader autograder/`
+
+2. Run the grader locally using [`coursera_autograder`](https://github.com/coursera/coursera_autograder) and provided sample submissions in the `sample-submissions` folder:
+<br>`$ coursera_autograder grade local pythongrader sample-submissions/sample-flipSign/ '{"partId":"5ShhY", "fileName":"flipSign.py"}' --dst-dir ~/Desktop`
 <br>or
-<br>`$ courseraprogramming grade local autograder sample-submissions/sample-Multiply/ partId Jn4Lh`
+<br>`$ $ coursera_autograder grade local pythongrader sample-submissions/sample-multiply/ '{"partId":"Zb6wb", "fileName":"multiply.py"}' --dst-dir ~/Desktop`
 
 3. Once you've tried testing locally, it's now time to upload the grader to the platform. First, create a Programming Assignment item on the platform, and update the `partId`s in `testCases.py` to match the Programming Assignment item. If you do not modify this, you will receive 0 points and `partId` errors upon submission. Then, you can upload the build files or the built Docker image to the Coursera platform. If you're uploading the build files, create a .zip file of all files in the `/autograder` folder.
