+++ custom-graders/v1/DemoJavaGrader/GraderFiles/Dockerfile
